Rename file `tap-page` to `tab-page`
